split the catalog linter out of the unit test suite
schema.test.ts iterated all 180 content/wads/*.json files with ~8
assertions each, so 'pnpm test' reported ~1,450 tests of which 37 are
code tests — the rest is content validation, including 6 live YouTube
oEmbed requests that could fail the suite offline.

Catalog checks now live in content/catalog.test.ts: 'pnpm test' runs
code tests (src), 'pnpm test:content' lints the catalog (YouTube
existence checks opt-in via CHECK_YOUTUBE=1), 'pnpm test:all' runs both.
schema.test.ts keeps only the schema-rule unit tests.
